Minor usage tweak
Entering a code in configMode now toggles whether the button is active rather than just setting it to true. This allows you to turn off buttons that may have entered by error or were entered by recalling the last set of buttons. Done by re-entering the code for the button.